Install ruby 1.9.2 if missing and rvm *is* already present

+++ bootstrap.sh
@@ -13,7 +13,7 @@ WORK_DIR=/tmp/lunar-station-`date +%s`
 mkdir -p $WORK_DIR
 cd $WORK_DIR
 
-if [ $OSTYPE == "linux-gnu" ]; then
+if [ $OSTYPE = "linux-gnu" ]; then
   OS="linux"
   if [ -x /usr/bin/apt-get ]; then
     PLATFORM="ubuntu"
@@ -52,7 +52,9 @@ if [ ! -d ~/.rvm ]; then
   log "Appending RVM function to your .bash_profile..."
   grep "scripts/rvm" ~/.bash_profile >/dev/null 2>&1 || echo '[[ -s "$HOME/.rvm/scripts/rvm" ]] && . "$HOME/.rvm/scripts/rvm" # Load RVM function' >> ~/.bash_profile
   . ~/.rvm/scripts/rvm
+fi
 
+if [[ ! $(rvm list | grep -F 1.9.2) ]]; then
   log "Installing ruby 1.9.2..."
   if [ $PLATFORM = "ubuntu" ]; then
     sudo apt-get -y install build-essential bison openssl libreadline6 libreadline6-dev zlib1g zlib1g-dev libssl-dev libyaml-dev libsqlite3-0 libsqlite3-dev sqlite3 libxml2-dev libxslt-dev autoconf libc6-dev ncurses-dev
